feat(#1868): PIPE/STREAM entry_type in SysReadResult + dispatch_adapters.rs split
- SysReadResult.entry_type: DT_PIPE(3)/DT_STREAM(4) for IPC dispatch,
  0 for normal miss. Python can dispatch without redundant detection.
- dispatch_adapters.rs: new file for Direction 2 (DISPATCH) PyO3 adapters
  (PyInterceptHookAdapter, PyPathResolverAdapter, PyMutationObserverAdapter).
  Separated from generated_adapters.rs (which is codegen-managed).
- dispatch.rs: pure Rust ABI (traits + registries, no PyO3 adapters)
- TODO: autogen dispatch_adapters.rs via codegen (next PR)
